Build release binaries on github actions
Adds `Package` and `Publish` steps to the main GitHub Actions workflow
to build release archives and publish them to the releases page whenever
a tag is pushed to the repository.

The `bin/package` script builds the archives, and `bin/install` provides
a  handy way to install those archives.

Also adds instructions for how to use the install script to the readme.